requirements.txt
faiss-cpu
langchain
openai==0.28
tiktoken
python-dotenv
fastapi
uvicorn
pydantic
